Update to Minecraft 1.12.2

+++ proxy/src/main/java/net/md_5/bungee/UserConnection.java
@@ -91,7 +91,7 @@ public final class UserConnection implements ProxiedPlayer
     /*========================================================================*/
     @Getter
     @Setter
-    private int sentPingId;
+    private long sentPingId;
     @Getter
     @Setter
     private long sentPingTime;
